refactor(Scanner): remove finalize method and leave close of AutoCloseable
The finalize method is usually not recommended as a way of clearing
resources. So much so that it has been deprecated in Java 9. Instead,
other ways are recommededed. The AutoCloseable (and the Closeable)
interfaces are one way of handling this, particularly when working with
the try-with statement, which autocloses any resource that implements
the AutoCloseable and Closeable interfaces. Class Buffer has been
refactored to follow this practice.