refactor: create new package structure
The original Coco/R for Java, which was translated from the C# version,
has all classes concentrated in a single package named Coco. This was
probably how the C# version has been implemented from one of the Pascal
sources. The original Coco/R software was implemented in Oberon, and
then a port to Modula-2, Pascal, C++, and a few other languages probably
followed the Oberon structure. Oberon's modular structure is not as
sophisticated as Java's or C#'s. As such, it was common to have all
files in a single folder. Java offers a better way to organize our code
through packages, and since Java 9, through modules as well. In order to
take advantage of this feature, the original source code has been
restructured into the following packages:

- application for the application code
- dfa for all dfa related classes
- generators for all code generators, but for now only parser generator
- parser for all parser related classes
- scanner for all the classes related to scanning
- source for all classes related to source code processing; will
  probably be split into source and io
- symbol for all classes related to the symbols returned by the scanner
- table for all symbol table processing
- trace for source code tracing and statistics; will probably be renamed
  to stats in the future